+++ lib/color.c
@@ -2,7 +2,7 @@
 #include "./util.h"
 #include <stdio.h>
 
-Color color(float r, float g, float b) { return (Color){r, g, b}; }
+Color color(double r, double g, double b) { return (Color){r, g, b}; }
 Color color_add(Color a, Color b) {
   return color(a.red + b.red, a.green + b.green, a.blue + b.blue);
 }
@@ -12,10 +12,10 @@ Color color_sub(Color a, Color b) {
 Color color_mul(Color a, Color b) {
   return color(a.red * b.red, a.green * b.green, a.blue * b.blue);
 }
-Color color_smul(Color a, float by) {
+Color color_smul(Color a, double by) {
   return color(a.red * by, a.green * by, a.blue * by);
 }
-Color color_sdiv(Color a, float by) {
+Color color_sdiv(Color a, double by) {
   return color(a.blue / by, a.green / by, a.blue / by);
 }
 bool color_equal(Color a, Color b) {
